+++ app/exceptions/sm_sms_campaign_webhook/error.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module SmSmsCampaignWebhook
+  # General base error type for custom errors.
+  class Error < StandardError; end
+end