+++ app/exceptions/sm_sms_campaign_webhook/invalid_payload_value.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module SmSmsCampaignWebhook
+  # Error type for invalid value in the payload received by the SMS webhook.
+  class InvalidPayloadValue < Error; end
+end