+++ lib/sm_sms_campaign_webhook.rb
@@ -5,6 +5,4 @@ require "sm_sms_campaign_webhook/version"
 
 # Namespace for SMS campaign webhook.
 module SmSmsCampaignWebhook
-  # General base error type for custom errors.
-  class Error < StandardError; end
 end